UI method class, tile hover effect, sorted buttons, resource manager
UI methods class is a wrapper that will hold all the UI methods for now, this is way cleaner than putting UI methods in each class separately given the limitations of what the UI can actually call.

UI is now a bit cleaner so I can debug buttons more easily.

Added a simple particle system when hovering over a tile that is self contained and listens to the tile hovered event.

Added a class that loads resources from the /resources folder. Haven't decided exactly how I want to go about sorting my prefabs so its very simple (and bad) right now but gets the job done. The job is that now class can get a resource without having to have a reference to it in their own class, they only need to know the ID. Eventually there will be an enum with types of objects or something and then very few classes will need to touch prefabs or file paths at all.