UnitRegistry, TurnHandler and GridController -> GameComponents
Changed UnitRegistry, TurnHandler and GridController from singletons to managed instances in the GameComponents class.

Clean up of units is now done in the respective classes when UnitDestroyed event is fired. Will probably move the unit setup to an event based thing as well.
What I should probably do is make a third class that handles the events for Unit being created/destroyed and then calls methods from the involved classes.